% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryWtd.R
\name{coef,wtd-method}
\alias{coef,wtd-method}
\title{Extract coefficients}
\usage{
\S4method{coef}{wtd}(object)
}
\arguments{
\item{object}{a fitted object of class inheriting from "wtd"}
}
\value{
A summary of the fitted object
}
\description{
Extract the coefficients from an object of class inheriting from "wtd"
}
