% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfunctions.R
\name{dlnorm}
\alias{dlnorm}
\title{The Lognormal distribution}
\usage{
dlnorm(x, logitp, mu, lnsigma, delta = 1, log = FALSE)
}
\arguments{
\item{x}{vector of prescription redemption times}

\item{logitp}{log-odds of being a prevalent user}

\item{mu}{mean on log-scale}

\item{lnsigma}{log of standard deviation on log-scale}

\item{delta}{XXXX}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
The Lognormal distribution
}
