% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_class.R
\name{predict,wtd-method}
\alias{predict,wtd-method}
\title{Predict Method for wtd Fits (probability or duration)}
\usage{
\S4method{predict}{wtd}(
  object,
  newdata = NULL,
  type = "dur",
  distrx = NULL,
  quantile = 0.8,
  se.fit = FALSE,
  na.action = na.pass,
  ...
)
}
\arguments{
\item{wtd}{a fitted object of class inheriting from "wtd"}
}
\value{
A vector of predictions
}
\description{
Predict Method for wtd Fits (probability or duration)
}
