% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranwtdttt.R
\name{ranwtdttt}
\alias{ranwtdttt}
\title{Fit Waiting Time Distribution with random index times}
\usage{
ranwtdttt(
  form,
  parameters = NULL,
  data,
  id,
  start,
  end,
  reverse = F,
  nsamp = 4,
  subset,
  na.action = na.pass,
  init,
  control = NULL,
  ...
)
}
\arguments{
\item{form}{an object of class "formula" (or one that can be coerced to that
class): a symbolic description of the model to be fitted. The details of the
model specification are given under 'Details'}

\item{parameters}{model formulae for distribution parameters}

\item{data}{an optional data frame, list or environment (or object coercible
by as.data.frame to a data frame) containing the variables in the model. If
not found in data, the variables are taken from environment(formula),
typically the environment from which wtdttt is called.}

\item{id}{the name of the variable that identifies distinct individuals}

\item{start}{start of observation window}

\item{end}{end of observation window}

\item{reverse}{logical; Fit the reverse waiting time distribution.}

\item{nsamp}{number of samples to take.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{na.action}{a function which indicates what should happen when the data
contain NAs. The default is set by the na.action setting of options, and is
na.fail if that is unset. The 'factory-fresh' default is na.omit. Another
possible value is NULL, no action. Value na.exclude can be useful.}

\item{init}{starting values for the parameters.}

\item{control}{a list of parameters for controlling the fitting process.}

\item{...}{further arguments passed to other methods.}
}
\value{
wtdttt returns an object of class "wtd" inheriting from "mle".
}
\description{
\code{ranwtdttt()} estimates maximum likelihood estimates for parametric Waiting Time Distribution (WTD)
based on observed prescription redemptions with adjustment for covariates
using one or more random index times for each individual. It reports estimates
of prevalence fraction and specified percentile of inter-arrival density
together with regression coefficients.
}
