% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rwtdttt-package.R
\docType{package}
\name{Rwtdttt-package}
\alias{Rwtdttt}
\alias{Rwtdttt-package}
\title{Rwtdttt: Parametric Waiting Time Distribution estimation}
\description{
Estimation of prescription durations and treatment probability based on the parametric Waiting Time Distribution. Pharmacoepidemiologic databases contains information on medication dispensings at pharmacies. Studies using such data typically require some estimate of duration of treatment after a dispensing (known as the prescription duration), which can be estimated using the parametric Waiting Time Distribution.
}
\keyword{internal}
