% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{sand_vcov}
\alias{sand_vcov}
\title{Calculate a robust variance-covariance matrix using the sandwich estimator}
\usage{
sand_vcov(fit)
}
\arguments{
\item{fit}{an object of class "wtd" returned by \code{ranwtdttt()}}
}
\value{
sand_vcov returns a matrix
}
\description{
Calculate a robust variance-covariance matrix using the sandwich estimator
}
