% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryWtd.R
\name{show,summary.wtd-method}
\alias{show,summary.wtd-method}
\title{Formatting of the summary of the model object and the prevalence of drug use}
\usage{
\S4method{show}{summary.wtd}(object)
}
\arguments{
\item{object}{a fitted object of class inheriting from "wtd"}
}
\value{
A summary of the fitted object along with the estimated prevalence and its 95\% CI
}
\description{
Formatting of the summary of the model object and the prevalence of drug use
}
