% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryWtd.R
\name{summary,wtd-method}
\alias{summary,wtd-method}
\title{Summarise model object and prevalence of drug use}
\usage{
\S4method{summary}{wtd}(object)
}
\arguments{
\item{object}{a fitted object of class inheriting from "wtd"}
}
\value{
A summary of the fitted object
}
\description{
Summarise model object from wtdttt and print the estimated prevalence of drug use along
with its 95\% confidence interval .
}
