% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_wtd-class.R
\docType{class}
\name{summary.wtd-class}
\alias{summary.wtd-class}
\title{summary.wtd-class - R class definition}
\description{
Register a 'summary.wtd' class, inheriting from 'summary.mle2'
}
\section{Slots}{

\describe{
\item{\code{call}}{call of the model formula}

\item{\code{coef}}{matrix of coefficients estimated through wtdttt function}

\item{\code{m2logL}}{log-likelihood}

\item{\code{prev_fin}}{estimate of prevalence of drug use and its 95\% confidence interval}
}}

