% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryWtd.R
\name{vcov,wtd-method}
\alias{vcov,wtd-method}
\title{Extract variance-covariance matrix}
\usage{
\S4method{vcov}{wtd}(object)
}
\arguments{
\item{object}{a fitted object of class inheriting from "wtd"}
}
\value{
A vector of coefficients
}
\description{
Extract variance-covariance matrix from an object of class inheriting from "wtd"
}
