% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd-class.R
\docType{class}
\name{wtd-class}
\alias{wtd-class}
\title{An S4 class representing the result of maximum likelihood estimation for a WTD}
\description{
This class encapsulates results of maximum likelihood estimateion for a
waiting time distribution.
}
\section{Slots}{

\describe{
\item{\code{delta}}{value of the delta parameter}

\item{\code{dist}}{name of the WTD distribution family: "lnorm", "weib" or "exp"}

\item{\code{depvar}}{name of the dependent variable}

\item{\code{idvar}}{(optional) name of the id variable}
}}

