% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtdttt.R
\name{wtdttt}
\alias{wtdttt}
\title{Fit Waiting Time Distribution}
\usage{
wtdttt(
  data,
  form,
  parameters = NULL,
  start = NA,
  end = NA,
  reverse = F,
  id = NA,
  subset = NA,
  na.action = na.pass,
  init = NULL,
  control = NULL,
  ...
)
}
\arguments{
\item{data}{an optional data frame, list or environment (or object coercible
by as.data.frame to a data frame) containing the variables in the model. If
not found in data, the variables are taken from environment(formula),
typically the environment from which wtdttt is called.}

\item{form}{an object of class "formula" (or one that can be coered to that
class): a symbolic description of the model to be fitted. The details of the
model specification are given under 'Details'}

\item{parameters}{optional model formulae for distribution parameters}

\item{start}{start of observation window (date or real number)}

\item{end}{end of observation window (date or real number)}

\item{reverse}{logical; Fit the reverse waiting time distribution (default F).}

\item{id}{name of the id variable (optional)}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{na.action}{a function which indicates what should happen when the data
contain NAs. The default is set by the na.action setting of options, and is
na.fail if that is unset. The 'factory-fresh' default is na.omit. Another
possible value is NULL, no action. Value na.exclude can be useful.}

\item{init}{starting values for the parameters.}

\item{control}{a list of parameters for controlling the fitting process.}

\item{...}{further arguments passed to other methods.}
}
\value{
wtdttt returns an object of class "wtd" inheriting from "mle".
}
\description{
Estimates the maximum likelihood estimate for a parametric Waiting Time
Distribution (WTD) based on observed prescription redemptions with
adjustment for covariates. Reports estimates of prevalence fraction
and specified percentile of inter-arrival density together with
regression coefficients.
}
\section{Model formula}{

The model formula \code{form} follows the pattern \code{obstime ~ dist(alpha, beta, gamma)}
with
\itemize{
\item \code{obstime}: the redemption time variable (date or real number)
\item \code{dist}: the parametric distribution for the forward or backward recurrence
density (FRD/BRD), which must be \code{dexp()}, \code{dweib()} or \code{dlnorm()}
i.e named after their corresponding interarrival density (IAD).
}
}

\section{Data format}{

The WTD is fit to the first prescription redemption of each
individual within an observation window (ordinary WTD), or the last
(reverse WTD), respectively.

You may prepare the data to this format, or optionally specify the name of
an \code{id} variable to select the first or last redemption automatically.

If the redemption time data are of type date, a continuity correction will be
applied automatically.
}

