% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfunctions.R
\name{dexp}
\alias{dexp}
\title{The Exponential distribution}
\usage{
dexp(x, logitp, lnbeta, delta = 1, log = FALSE)
}
\arguments{
\item{x}{vector of event times (must be between (0; delta))}

\item{logitp}{log-odds of being a prevalent user}

\item{lnbeta}{log of beta (scale)}

\item{delta}{width of interval with positive support (x in (0; delta))}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}
}
\value{
Density. The length of the result is the maximum of the lengths of the numerical arguments.
}
\description{
The Exponential distribution
}
