% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_class.R
\name{dlnorm}
\alias{dlnorm}
\title{The Lognormal Distribution}
\usage{
dlnorm(x, logitp, mu, lnsigma, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{logitp}{how to describe this?}

\item{mu}{mean}

\item{lnsigma}{log of standard deviation}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
The Lognormal Distribution
}
