% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfunctions.R
\name{dlnorm}
\alias{dlnorm}
\title{The Lognormal distribution}
\usage{
dlnorm(x, logitp, mu, lnsigma, delta = 1, log = FALSE)
}
\arguments{
\item{x}{vector of event times (must be between (0; delta))}

\item{logitp}{log-odds of being a prevalent user}

\item{mu}{mean on log-scale}

\item{lnsigma}{log of standard deviation on log-scale}

\item{delta}{width of interval with positive support (x in (0; delta))}

\item{log}{logical; if TRUE, density values are returned on log-scale.}
}
\value{
Density. The length of the result is the maximum of the lengths of the numerical arguments.
}
\description{
The Lognormal distribution
}
