% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,wtd,ANY-method}
\alias{plot,wtd,ANY-method}
\title{Make WTD diagnostic plots}
\usage{
\S4method{plot}{wtd,ANY}(wtd, x = NULL, y = NULL, ...)
}
\arguments{
\item{wtd}{wtd object, typically result of wtdttt}

\item{x}{ignored}

\item{y}{ignored}

\item{...}{other graphical parameters (see \code{par})}
}
\description{
Make diagnostic plots showing the fit of an estimated parametric Waiting Time
Distribution (WTD) with respect to the observed histogram of prescription
redemptions.
}
