% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,wtd,ANY-method}
\alias{plot,wtd,ANY-method}
\title{Make WTD diagnostic plots}
\usage{
\S4method{plot}{wtd,ANY}(object, x, y, ...)
}
\arguments{
\item{wtd}{wtd object, typically result of wtdttt}
}
\description{
Make diagnostic plots showing the fit of an estimated parametric Waiting Time
Distribution (WTD) with respect to the observed histogram of prescription
redemptions.
HS 240108: I think this needs to incorporate information on time period
of observations (start and end specified in call to wtdttt)
}
