% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_class.R
\name{plot,wtd,ANY-method}
\alias{plot,wtd,ANY-method}
\title{Plot Diagnostics for a wtd Object}
\usage{
\S4method{plot}{wtd,ANY}(object, x, y, ...)
}
\arguments{
\item{wtd}{wtd object, typically result of wtdttt}
}
\description{
Plot Diagnostics for a wtd Object
}
