% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_dur_prob.R
\name{predict,wtd-method}
\alias{predict,wtd-method}
\title{Make WTD predictions}
\usage{
\S4method{predict}{wtd}(
  object,
  prediction.data = NULL,
  type = "dur",
  iadmean = F,
  distrx = NULL,
  quantile = 0.8,
  se.fit = FALSE,
  na.action = na.pass,
  ...
)
}
\arguments{
\item{object}{a fitted object of class inheriting from "wtd"}

\item{prediction.data}{An optional data frame in which to look for variables with
which to predict. If omitted, the fitted values are used.}

\item{type}{"dur" or "prob". Default "dur".}

\item{iadmean}{logical; if T, mean duration is predicted.}

\item{distrx}{For type="prob", a vector of dispensing gaps.}

\item{quantile}{For type="dur", quantile of distribution. Default 0.8}

\item{se.fit}{A switch indicating if standard errors are required}

\item{na.action}{function determining what should be done with missing values in
\code{newdata}. The default is to predict \code{NA}}

\item{...}{further arguments passed to or from other methods}
}
\value{
A vector of predictions
}
\description{
Make predictions based on an estimated parametric Waiting Time Distribution
(WTD) model, either the probability of a person still being in treatment or
the duration of observed prescription redemptions.
}
\details{
#' @section Data format:
Covariates in the data provided for wtdttt, used for estimation, must be in the prediction data as well -if used- with identical names
The class of the covariate must be the same in the dataset used for estimation and in the one used for prediction - if different - i.e. packsize used as a factor.
}
